Move the Social Profile Helper to the right folder and add it to the surface
